refactor(threading): Rewrite HumanPoseEstimation using AsyncJob
This fixes thread-safety and prevents crashes
